requirements.txt
streamlit
pandas
python-dotenv
pdfplumber
openpyxl
python-dotenv
langchain
langchain_community
langchain_core
langchain_chroma
huggingface_hub
sentence_transformers
unstructured
langchain_community
langchain_huggingface
langchain_openai
langchain_groq
rapidfuzz